+++ packages/components-react/src/pages/ContentWrapper.tsx
@@ -32,6 +32,26 @@ export const ContentWrapperPage = (): JSX.Element => {
         </PContentWrapper>
       </div>
 
+      <div
+        className="playground"
+        title="should render with 'default' background color"
+        style={{ background: 'deeppink' }}
+      >
+        <PContentWrapper backgroundColor="default">
+          <p>Some content</p>
+        </PContentWrapper>
+      </div>
+
+      <div
+        className="playground"
+        title="should render with 'default' background color and dark theme"
+        style={{ background: 'deeppink' }}
+      >
+        <PContentWrapper backgroundColor="default" theme="dark">
+          <p>Some content</p>
+        </PContentWrapper>
+      </div>
+
       <div className="playground" title="should color full width although width 'basic' is enabled">
         <PContentWrapper width="basic" style={{ background: 'deeppink' }}>
           <p>Some content</p>