+++ packages/components/CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### [Unreleased]
 
-#### Changed
+#### Added
+- Configurable background color of `Content Wrapper`
 - `italic` font-style in `Text` is now overridden with `normal`
 
 #### Fixed